Fixed: Function AddContextEntry not working correctly on HS clients >= 7.0.0.0 Fixed: Dead bonded pets not updating the ghost corpse correctly on resurrect. Fixed: Explosion potions causing damage to nearby items when it explodes.
Internal changes:
-Fixed graysvr.h calling undefined SocketsReceive()
-Optimized EF_ItemStacking / EF_ItemStackDrop code. Although it's usable, it still have some minor problems as before. Also moved EF_ItemStacking from CItem::MoveToCheck() to CChar::ItemDrop() to match EF_ItemStackDrop located on CChar::ItemPickup()

+++ 0.56.b/network/send.cpp
@@ -3872,7 +3872,7 @@ void PacketDisplayPopup::addOption(WORD entryTag, DWORD textId, WORD flags, WORD
 		if (flags & POPUPFLAG_COLOR)
 			flags &= ~POPUPFLAG_COLOR;
 
-		writeInt32(static_cast<WORD>(textId));
+		writeInt32(textId);
 		writeInt16(entryTag);
 		writeInt16(flags);
 	}
